rename OrderedNetwork to MessageQueue
